Added support for sqlite v3, sqlite3hash.rb and sqlite3store.rb.. Refactored state machines in account to multiple routines and added   comments.. Fixed order of message issues with disconnection, moved disconnect   to account. Add support to @edit command to edit World.msgs. Removed :logged_out event in favor of one :disconnected event. Changed :logged_out event issued from Engine side to :quit to prevent  confusion.

+++ dbload.rb
@@ -32,7 +32,7 @@ class Loader
   VERSION = "0.2.0"
 
   attr_accessor :opts
-  DATABASES = [:dbm, :gdbm, :sdbm, :sqlite]
+  DATABASES = [:dbm, :gdbm, :sdbm, :sqlite, :sqlite3]
 
   def initialize
     @opts = get_options
@@ -46,6 +46,9 @@ class Loader
     when :sqlite
       require 'sqlite'
       require 'storage/sqlitehash'
+    when :sqlite3
+      require 'sqlite3'
+      require 'storage/sqlite3hash'
     end
     @dbtop = 0
     @db = {}
@@ -96,6 +99,7 @@ class Loader
     myopts.ofile = myopts.ifile.dup if myopts.ofile.nil?
     myopts.ofile << ".gdbm" if myopts.dbtype == :gdbm
     myopts.ofile << ".sqlite" if myopts.dbtype == :sqlite
+    myopts.ofile << ".sqlite3" if myopts.dbtype == :sqlite3
     myopts.ifile << ".yaml"
 
     return myopts
@@ -136,8 +140,17 @@ class Loader
         db.execute("drop table tmud;")
       rescue
       end
-      db.execute("create table tmud (id integer unique, data text);")
-      @db.each {|k,v| db[k.to_s] = Utility.encode(v)}
+      db.execute("create table tmud (id integer primary key, data text);")
+      @db.each {|k,v| db[k] = Utility.encode(v)}
+      db.close
+    when :sqlite3
+      db = SQLite3::Database.open(@opts.ofile)
+      begin
+        db.execute("drop table tmud;")
+      rescue Exception
+      end
+      db.execute("create table tmud (id integer primary key, data text);")
+      @db.each {|k,v| db[k] = Utility.encode(v)}
       db.close
     end
